Make it possible to override STICKY on class level.
The list of sticky groups/members on the sharing view was defined at module
level, but also at the class level. The idea was likely that this can be
easily overriden by extending this view and only changing the class-level
STICKY member, but because the rest of the code referred to the module-level
variable, this would not work. This two-liner change should fix that.